Fixes to dmu engine and ability to configure fio with ZoL repo
New configure option --with-zfs= to build fio with prebuilt ZoL repository.
The DMU engine still does not fully work as it fails to open exported pool
which was created by kernel ZFS on linux.

This commit is part of effort described in following ticket:
CCS-71 As a developer I want to test iops of uZFS using FIO